pyproject.toml
[tool.poetry]
name = "bloom-scraper"
version = "0.1.0"
description = ""
authors = ["Tim Sharpe <tim@bombasticmonkey.com>"]
readme = "README.md"
packages = [{include = "bloom_scraper"}]

[tool.poetry.scripts]
bloom_scraper = "bloom_scraper.cli:main"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.18.1"
webdriver-manager = "^4.0.1"
beautifulsoup4 = "^4.12.3"
tenacity = "^8.2.3"
pypdf2 = "^3.0.1"
googletrans-py = "^4.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
